Fix #128121
