+++ src/vs/workbench/services/extensionManagement/common/webExtensionManagementService.ts
@@ -4,46 +4,30 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ExtensionType, IExtensionIdentifier, IExtensionManifest } from 'vs/platform/extensions/common/extensions';
-import { IExtensionManagementService, ILocalExtension, InstallExtensionEvent, DidUninstallExtensionEvent, IGalleryExtension, IReportedExtension, IGalleryMetadata, InstallOperation, IExtensionGalleryService, ExtensionManagementError, INSTALL_ERROR_INCOMPATIBLE, InstallOptions, InstallExtensionResult } from 'vs/platform/extensionManagement/common/extensionManagement';
-import { Event, Emitter } from 'vs/base/common/event';
+import { IExtensionManagementService, ILocalExtension, IGalleryExtension, IGalleryMetadata, InstallOperation, IExtensionGalleryService, InstallOptions } from 'vs/platform/extensionManagement/common/extensionManagement';
 import { URI } from 'vs/base/common/uri';
 import { areSameExtensions, getGalleryExtensionId } from 'vs/platform/extensionManagement/common/extensionManagementUtil';
 import { IScannedExtension, IWebExtensionsScannerService } from 'vs/workbench/services/extensionManagement/common/extensionManagement';
 import { ILogService } from 'vs/platform/log/common/log';
-import { Disposable } from 'vs/base/common/lifecycle';
-import { localize } from 'vs/nls';
 import { IExtensionManifestPropertiesService } from 'vs/workbench/services/extensions/common/extensionManifestPropertiesService';
 import { CancellationToken } from 'vs/base/common/cancellation';
-import { IProductService } from 'vs/platform/product/common/productService';
+import { AbstractExtensionManagementService, AbstractExtensionTask, IInstallExtensionTask, IUninstallExtensionTask, UninstallExtensionTaskOptions } from 'vs/platform/extensionManagement/common/abstractExtensionManagementService';
+import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 
-type Metadata = {
-	isMachineScoped?: boolean;
-};
+type Metadata = Partial<IGalleryMetadata & { isMachineScoped: boolean; }>;
 
-export class WebExtensionManagementService extends Disposable implements IExtensionManagementService {
+export class WebExtensionManagementService extends AbstractExtensionManagementService implements IExtensionManagementService {
 
 	declare readonly _serviceBrand: undefined;
 
-	private readonly _onInstallExtension = this._register(new Emitter<InstallExtensionEvent>());
-	readonly onInstallExtension: Event<InstallExtensionEvent> = this._onInstallExtension.event;
-
-	private readonly _onDidInstallExtension = this._register(new Emitter<InstallExtensionResult[]>());
-	readonly onDidInstallExtensions = this._onDidInstallExtension.event;
-
-	private readonly _onUninstallExtension = this._register(new Emitter<IExtensionIdentifier>());
-	readonly onUninstallExtension: Event<IExtensionIdentifier> = this._onUninstallExtension.event;
-
-	private _onDidUninstallExtension = this._register(new Emitter<DidUninstallExtensionEvent>());
-	onDidUninstallExtension: Event<DidUninstallExtensionEvent> = this._onDidUninstallExtension.event;
-
 	constructor(
+		@IExtensionGalleryService extensionGalleryService: IExtensionGalleryService,
+		@ITelemetryService telemetryService: ITelemetryService,
+		@ILogService logService: ILogService,
 		@IWebExtensionsScannerService private readonly webExtensionsScannerService: IWebExtensionsScannerService,
-		@ILogService private readonly logService: ILogService,
-		@IExtensionGalleryService private readonly extensionGalleryService: IExtensionGalleryService,
 		@IExtensionManifestPropertiesService private readonly extensionManifestPropertiesService: IExtensionManifestPropertiesService,
-		@IProductService private readonly productService: IProductService,
 	) {
-		super();
+		super(extensionGalleryService, telemetryService, logService);
 	}
 
 	async getInstalled(type?: ExtensionType): Promise<ILocalExtension[]> {
@@ -56,15 +40,15 @@ export class WebExtensionManagementService extends Disposable implements IExtens
 			const userExtensions = await this.webExtensionsScannerService.scanUserExtensions();
 			extensions.push(...userExtensions);
 		}
-		return Promise.all(extensions.map(e => this.toLocalExtension(e)));
+		return Promise.all(extensions.map(e => toLocalExtension(e)));
 	}
 
 	async canInstall(gallery: IGalleryExtension): Promise<boolean> {
-		const compatibleExtension = await this.extensionGalleryService.getCompatibleExtension(gallery);
+		const compatibleExtension = await this.galleryService.getCompatibleExtension(gallery);
 		if (!compatibleExtension) {
 			return false;
 		}
-		const manifest = await this.extensionGalleryService.getManifest(compatibleExtension, CancellationToken.None);
+		const manifest = await this.galleryService.getManifest(compatibleExtension, CancellationToken.None);
 		if (!manifest) {
 			return false;
 		}
@@ -74,90 +58,99 @@ export class WebExtensionManagementService extends Disposable implements IExtens
 		return true;
 	}
 
-	async install(location: URI, installOptions?: InstallOptions): Promise<ILocalExtension> {
+	async install(location: URI, options: InstallOptions = {}): Promise<ILocalExtension> {
+		this.logService.trace('ExtensionManagementService#install', location.toString());
 		const manifest = await this.webExtensionsScannerService.scanExtensionManifest(location);
 		if (!manifest) {
 			throw new Error(`Cannot find packageJSON from the location ${location.toString()}`);
 		}
-
-		const identifier = { id: getGalleryExtensionId(manifest.publisher, manifest.name) };
-		this.logService.info('Installing extension:', identifier.id);
-		this._onInstallExtension.fire({ identifier: identifier, source: location.path });
-
-		try {
-			const userExtensions = await this.webExtensionsScannerService.scanUserExtensions();
-			const existingExtension = userExtensions.find(e => areSameExtensions(e.identifier, identifier));
-			const metadata = this.getMetadata(installOptions, existingExtension);
-
-			const extension = await this.webExtensionsScannerService.addExtension(location, metadata);
-			const local = this.toLocalExtension(extension);
-			this._onDidInstallExtension.fire([{ local, identifier: extension.identifier, operation: InstallOperation.Install }]);
-			return local;
-		} catch (error) {
-			this._onDidInstallExtension.fire([{ identifier, operation: InstallOperation.Install }]);
-			throw error;
-		}
-	}
-
-	async installFromGallery(gallery: IGalleryExtension, installOptions?: InstallOptions): Promise<ILocalExtension> {
-		this.logService.info('Installing extension:', gallery.identifier.id);
-		this._onInstallExtension.fire({ identifier: gallery.identifier, source: gallery });
-		try {
-			const compatibleExtension = await this.extensionGalleryService.getCompatibleExtension(gallery);
-			if (!compatibleExtension) {
-				throw new ExtensionManagementError(localize('notFoundCompatibleDependency', "Unable to install '{0}' extension because it is not compatible with the current version of VS Code (version {1}).", gallery.identifier.id, this.productService.version), INSTALL_ERROR_INCOMPATIBLE);
-			}
-			const userExtensions = await this.webExtensionsScannerService.scanUserExtensions();
-			const existingExtension = userExtensions.find(e => areSameExtensions(e.identifier, compatibleExtension.identifier));
-			const metadata = this.getMetadata(installOptions, existingExtension);
-
-			const scannedExtension = await this.webExtensionsScannerService.addExtensionFromGallery(compatibleExtension, metadata);
-			const local = this.toLocalExtension(scannedExtension);
-			this._onDidInstallExtension.fire([{ local, identifier: compatibleExtension.identifier, operation: InstallOperation.Install, source: compatibleExtension }]);
-			return local;
-		} catch (error) {
-			this._onDidInstallExtension.fire([{ identifier: gallery.identifier, operation: InstallOperation.Install, source: gallery }]);
-			throw error;
-		}
-	}
-
-	async uninstall(extension: ILocalExtension): Promise<void> {
-		this._onUninstallExtension.fire(extension.identifier);
-		try {
-			await this.webExtensionsScannerService.removeExtension(extension.identifier);
-			this._onDidUninstallExtension.fire({ identifier: extension.identifier });
-		} catch (error) {
-			this.logService.error(error);
-			this._onDidUninstallExtension.fire({ error, identifier: extension.identifier });
-			throw error;
-		}
+		return this.installExtension(manifest, location, options);
 	}
 
 	async updateMetadata(local: ILocalExtension, metadata: IGalleryMetadata): Promise<ILocalExtension> {
 		return local;
 	}
 
-	private toLocalExtension(extension: IScannedExtension): ILocalExtension {
-		const metadata = this.getMetadata(undefined, extension);
-		return {
-			...extension,
-			isMachineScoped: !!metadata.isMachineScoped,
-			publisherId: null,
-			publisherDisplayName: null,
-		};
+	protected createInstallExtensionTask(manifest: IExtensionManifest, extension: URI | IGalleryExtension, options: InstallOptions): IInstallExtensionTask {
+		return new InstallExtensionTask(manifest, extension, options, this.webExtensionsScannerService);
 	}
 
-	private getMetadata(options?: InstallOptions, existingExtension?: IScannedExtension): Metadata {
-		const metadata: Metadata = {};
-		const existingExtensionMetadata: Metadata = existingExtension?.metadata || {};
-		metadata.isMachineScoped = options?.isMachineScoped || existingExtensionMetadata.isMachineScoped;
-		return metadata;
+	protected createUninstallExtensionTask(extension: ILocalExtension, options: UninstallExtensionTaskOptions): IUninstallExtensionTask {
+		return new UninstallExtensionTask(extension, options, this.webExtensionsScannerService);
 	}
 
 	zip(extension: ILocalExtension): Promise<URI> { throw new Error('unsupported'); }
 	unzip(zipLocation: URI): Promise<IExtensionIdentifier> { throw new Error('unsupported'); }
 	getManifest(vsix: URI): Promise<IExtensionManifest> { throw new Error('unsupported'); }
-	reinstallFromGallery(extension: ILocalExtension): Promise<void> { throw new Error('unsupported'); }
-	getExtensionsReport(): Promise<IReportedExtension[]> { throw new Error('unsupported'); }
 	updateExtensionScope(): Promise<ILocalExtension> { throw new Error('unsupported'); }
+}
+
+function toLocalExtension(extension: IScannedExtension): ILocalExtension {
+	const metadata = getMetadata(undefined, extension);
+	return {
+		...extension,
+		identifier: { id: extension.identifier.id, uuid: metadata.id },
+		isMachineScoped: !!metadata.isMachineScoped,
+		publisherId: metadata.publisherId || null,
+		publisherDisplayName: metadata.publisherDisplayName || null,
+	};
+}
+
+function getMetadata(options?: InstallOptions, existingExtension?: IScannedExtension): Metadata {
+	const metadata: Metadata = { ...(existingExtension?.metadata || {}) };
+	metadata.isMachineScoped = options?.isMachineScoped || metadata.isMachineScoped;
+	return metadata;
+}
+
+class InstallExtensionTask extends AbstractExtensionTask<ILocalExtension> implements IInstallExtensionTask {
+
+	readonly identifier: IExtensionIdentifier;
+	readonly source: URI | IGalleryExtension;
+	private _operation = InstallOperation.Install;
+	get operation() { return this._operation; }
+
+	constructor(
+		manifest: IExtensionManifest,
+		private readonly extension: URI | IGalleryExtension,
+		private readonly options: InstallOptions,
+		private readonly webExtensionsScannerService: IWebExtensionsScannerService,
+	) {
+		super();
+		this.identifier = URI.isUri(extension) ? { id: getGalleryExtensionId(manifest.publisher, manifest.name) } : extension.identifier;
+		this.source = extension;
+	}
+
+	protected async doRun(token: CancellationToken): Promise<ILocalExtension> {
+		const userExtensions = await this.webExtensionsScannerService.scanUserExtensions();
+		const existingExtension = userExtensions.find(e => areSameExtensions(e.identifier, this.identifier));
+		if (existingExtension) {
+			this._operation = InstallOperation.Update;
+		}
+
+		const metadata = getMetadata(this.options, existingExtension);
+		if (!URI.isUri(this.extension)) {
+			metadata.id = this.extension.identifier.uuid;
+			metadata.publisherDisplayName = this.extension.publisherDisplayName;
+			metadata.publisherId = this.extension.publisherId;
+		}
+
+		const scannedExtension = URI.isUri(this.extension) ? await this.webExtensionsScannerService.addExtension(this.extension, metadata)
+			: await this.webExtensionsScannerService.addExtensionFromGallery(this.extension, metadata);
+		return toLocalExtension(scannedExtension);
+	}
+}
+
+class UninstallExtensionTask extends AbstractExtensionTask<void> implements IUninstallExtensionTask {
+
+	constructor(
+		readonly extension: ILocalExtension,
+		options: UninstallExtensionTaskOptions,
+		private readonly webExtensionsScannerService: IWebExtensionsScannerService,
+	) {
+		super();
+	}
+
+	protected doRun(token: CancellationToken): Promise<void> {
+		return this.webExtensionsScannerService.removeExtension(this.extension.identifier);
+	}
 }
